Full update to 1.21.4 + changed the version to 1.0.2-1.21.4 Removed support for Player Data API (for now) Added Alt & Tab keys as ignored keys for AFK keyboard listener Fixed right-clicking Item Frame with Amethyst in both offhand and main hand causing the invisible (at the moment visible and glowing) item frame to become invisible again. Now this interaction is removed Fully moved away from rendering the navigation compass (instead of giving it as item) due to changes in Minecraft API Added manual config saving in many places (receiving TPC computation and AFK start/stop packets, navigation start/end) to avoid losing data when quitting Changed how WorldBlockPos is serialized and accessed from config due to regular implementation never working for some reason Added /qolmod, which contains essential info about mod features, and help messages for /svc, /skip(night|day) Slightly changed how Server Chest is stored and serialized; now Server Chest is saved every time any player exits its inventory (and when the server is stopped, as before) Now "/<command> help" also displays all aliases for the command (if they exist) /skip(night|day) inf now displays a proper message and preserves the previous number of skips that was set Skipping a night/day when duration is infinite but has a value of 0 no longer displays a warning about "skipping the last night/day" Added /skip(night|day) stop which is equivalent to /skip(night|day) 0 for more intuitive skip cancelling Added default constructor for WorldBlockPos that defaults to 0 0 0 in overworld
Took 6 hours 53 minutes

+++ src/main/kotlin/me/vannername/qol/client/utils/TPCreditsComputation.kt
@@ -56,7 +56,7 @@ object TPCreditsComputation {
         return abs(1 - distancesRelation) > 0.1
     }
 
-    fun tick(p: AbstractClientPlayerEntity, secondsBetweenTicks: Int) {
+    fun tick(p: AbstractClientPlayerEntity, secondsBetweenTicks: Double) {
         // only tick once per secondsBetweenTicks seconds
         if (System.currentTimeMillis() - lastTick < secondsBetweenTicks * 1000) {
             return
